Almost Final Version
Recreated the scene after last Unity project crashed and wouldn't load any of my old scripts. Reimported all my old assets and added more to the scene in terms of particle effects, new hand models for the controllers (figured that part out), and also fixed the color range of my MotionColor script so that it wouldn't stay on white majority of the time. Lastly, I also made some speaker models out of Unity prefabs and am about to start on creating instruments (sample-based) that the player can trigger when they hit the object with the keytar.